Move install type detection to manifest reader project
